feat(take): add higher-order lettable version of take
